TaconDeOro: Funciones añadir carrito y eliminar carrito

+++ src/tacondeoro/Articulo.java
@@ -71,7 +71,7 @@ public class Articulo {
         this.material = material;
     }
 
-    public double getPrecio() {
+    public float getPrecio() {
         return precio;
     }
